Befejezett abstract

+++ src/hu/flowacademy/ForceSensitive.java
@@ -3,6 +3,6 @@ package hu.flowacademy;
 public interface ForceSensitive {
     public boolean canDefeat(ForceSensitive forceSensitive);
 
-    public float howBigIsHisForce();
+    public double howBigIsHisForce();
 
 }
